requirements.txt
confusable-homoglyphs==3.2.0
Django==1.11
django-bootstrap4==0.0.8
django-friendship==1.8.1
django-registration==2.4.1
Pillow==6.0.0
psycopg2==2.8.2
pytz==2019.1
